xdp: add libxdp_clean_references
This change adds an API to libxdp to clean up dead references that
remain in bpffs after a net_device goes away, and the program is
detached. This helper can figure out if the pinned directory is dead,
and if so, clean it up.

Passing an ifindex of 0 has special semantics, in that it will traverse
through the entire bpffs tree looking for dead links, performing a
complete sweep.

This will be used to implement the xdp-loader clean command in a
subsequent patch.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>